requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pandas==1.3.3
requests==2.26.0
python-dotenv==0.19.0
gunicorn==20.1.0